+++ src/_assets/js/app.js
@@ -1,1 +1,3 @@
+import './../css/app.css';
+
 console.log( 'Hello world' );