+++ webpack.config.js
@@ -4,7 +4,8 @@ const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 module.exports = {
   
   name: process.env.NODE_ENV == 'production' ? 'production' : 'development',
-  mode: process.env.NODE_ENV == 'production' ? 'production' : 'development',
+  // mode: process.env.NODE_ENV == 'production' ? 'production' : 'development',
+  mode: 'production',
 
   entry: {
     'app': './src/_assets/js/app.js'