+++ webpack.config.js
@@ -4,8 +4,7 @@ const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 module.exports = {
   
   name: process.env.NODE_ENV == 'production' ? 'production' : 'development',
-  // mode: process.env.NODE_ENV == 'production' ? 'production' : 'development',
-  mode: 'production',
+  mode: process.env.NODE_ENV == 'production' ? 'production' : 'development',
 
   entry: {
     'app': './src/_assets/js/app.js'