+++ postcss.config.js
@@ -1,6 +1,7 @@
 module.exports = {
   plugins: [
     require('precss'), // Use Sass-like markup in your CSS
+    require('postcss-node-sass'), // Parse styles with node-sass
     require('postcss-preset-env'), // Convert modern CSS into something browsers understand
     require('autoprefixer'), // Add vendor prefixes
     require('cssnano')({ // Modern CSS compression