+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Install PostCSS
 - Install PostCSS plugins
 	- [x] autoprefixer
+	- [x] cssnano
 	- [x] postcss-preset-env
 	- [x] precss
 - Install webpack plugins