+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Install PostCSS plugins
 	- [x] autoprefixer
 	- [x] cssnano
+	- [x] postcss-node-sass
 	- [x] postcss-preset-env
 	- [x] precss
 - Install webpack plugins